refactor: move related modules to fake-bili
